Move more MatchObject pessimization to matcher
where it is used

CMK-14467
