+++ cmk/base/core_config.py
@@ -25,10 +25,10 @@ from cmk.utils.labels import Labels
 from cmk.utils.licensing.handler import LicenseState, LicensingHandler
 from cmk.utils.licensing.helper import get_licensed_state_file_path, write_licensed_state
 from cmk.utils.paths import core_helper_config_dir
-from cmk.utils.servicename import ServiceName
+from cmk.utils.servicename import Item, ServiceName
 from cmk.utils.store import load_object_from_file, lock_checkmk_configuration, save_object_to_file
 
-from cmk.checkengine.checking import CheckPluginName, ConfiguredService, Item, ServiceID
+from cmk.checkengine.checking import CheckPluginName, ConfiguredService, ServiceID
 from cmk.checkengine.parameters import TimespecificParameters
 
 import cmk.base.api.agent_based.register as agent_based_register
