+++ cmk/gui/wato/pages/pattern_editor.py
@@ -9,9 +9,9 @@ from collections.abc import Collection, Iterable
 
 from cmk.utils.hostaddress import HostName
 from cmk.utils.labels import Labels
-from cmk.utils.servicename import ServiceName
+from cmk.utils.servicename import Item, ServiceName
 
-from cmk.checkengine.checking import CheckPluginNameStr, Item
+from cmk.checkengine.checking import CheckPluginNameStr
 
 # Tolerate this for 1.6. Should be cleaned up in future versions,
 # e.g. by trying to move the common code to a common place
