+++ cmk/utils/servicename.py
@@ -3,6 +3,7 @@
 # This file is part of Checkmk (https://checkmk.com). It is subject to the terms and
 # conditions defined in the file COPYING, which is part of this source code package.
 
-__all__ = ["ServiceName"]
+__all__ = ["Item", "ServiceName"]
 
 ServiceName = str
+Item = str | None
